Try to avoid having Invent redirect to login or doing not found redirects
We can check the redirect on the server side. On Invent's side this is
another request, but it saves the user from making a request as well so
shouldn't add load on Invent.

Now, paths that aren't present would stay on this site to offer a chance
to edit the URL.